3.2.4 Base Stats Page, page num fix & erase A button if not caught on lore
if on page 3, then scroll up/down to next/prev mon, the base stats page doesnt land on the right page number, and is stuck on 1 past the normal max page. fixed by adding a const that automatically detects if there is max 3 pages (for using vanilla/StatEXP) or max of 4 pages (if you've replaced StatEXP with EVs) and then the reinitdexentry func in pokedex.asm uses this const to check and set the right page number.

also adds a func that simply erases the A button indicator and the page number if you hit the INFO (lore) button when you havent caught the mon yet. it now just shows the mon's species nickname

No longer displays dex lore/info, unless you've caught the pokemon. added UncaughtMon_Info_Erase_PageNum which erases the A button prompt and page num when this is the case

from my original CL commit:
fixed the base stats page num bug, added const in pokedex_stats_page.asm that sets the const value depending on if StatEXP or EVs are being used in the ROM, and more elegant solution in pokedex.asm